% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot the correlations between variables}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{ryx object.}

\item{...}{any other parameters passed to plot function}
}
\value{
a plot object
}
\description{
This code will take an ryx object and create a plot showing
the absolute value of correlation between the dependent and selected indpendent
variable
}
\examples{
\dontrun{
data(Boston, package = "MASS")
x <- ryx(Boston, y = "medv")
plot(x)
}
}
\seealso{
\link{ggplot}
}
