% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{prints correlation table}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{ryx object}

\item{...}{any other parameters passed to the print function}
}
\description{
This code will take an ryx object and print a table with the variable name,
correlations, p-value, and significance level between each of the variables
}
\examples{
\dontrun{
data(Boston, package = "MASS")
x <- ryx(Boston, y = "medv")
print(x)
}
}
