% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy.R
\name{ryx}
\alias{ryx}
\title{Correlation package that conducts tests between variables in a data set.}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{The data set to perform correlation tests on.}

\item{y}{The dependent variable.}

\item{x}{The independent variable.}
}
\value{
A list containing information about the correlation tests, p-values
and significance codes.
}
\description{
The function 'ryx' conducts correlation coefficients and their respective
p-values and significance codes between a dependent variable and specified
independent variables.
}
\examples{
data(mtcars)
result <- ryx(mtcars, y = "mpg", x = c("disp", "hp", "wt"))
print(result)
plot(result)
summary(result)
}
