% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summarize the correlations in a dataset}
\usage{
\method{summary}{ryx}(x, ...)
}
\arguments{
\item{x}{ryx object}

\item{...}{other parameters passed to the summary function}
}
\description{
This code will take an ryx object and summarize the findings
in text format
}
\examples{
\dontrun{
data(Boston, package = "MASS")
x <- ryx(Boston, y = "medv")
summary(x)
}
}
