% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot method for ryx objects}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class `ryx`, containing the correlation results from the `ryx` function.}
}
\description{
Creates a plot of the absolute correlations between the dependent variable and the independent variables.
}
\details{
The plot displays absolute correlation values, color-coded to indicate whether
the correlation is positive or negative. Positive correlations are shown in blue, and
negative correlations are shown in red. Independent variables are ordered by the magnitude
of their correlation with the dependent variable.
}
\examples{
library(MASS)
result <- ryx(Boston, y = "medv")
plot(result)

}
