% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot a ryx object}
\usage{
\method{plot}{ryx}(obj)
}
\arguments{
\item{obj}{A ryx object.}
}
\description{
Plots a ryx object using ggplot2.
}
\details{
The plot is arranged by the absolute value of each variable's correlation.
}
\examples{
library(MASS)
plot(ryx(Boston, y="medv"))
}
