% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{plot.ryx}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x:}{an object of class ryx}
}
\value{
A plot of the correlations of each variable.
}
\description{
The plot function displays the correlation of all the outcome variables against the response variable.
}
\examples{
\dontrun{
plot(ryx(mtcars, y = "mpg", x = c("hp", "wt", "disp", "cyl", "am", "gear")))
}
}
