% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{print.ryx}
\alias{print.ryx}
\title{Print method for ryx objects}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class `ryx`, containing the correlation results from the `ryx` function.}
}
\description{
Prints a table of correlation results for the dependent variable and each independent variable.
}
\details{
This method provides a concise table of correlations, p-values, and significance levels
between the dependent variable (`y`) and the independent variables (`x`).
}
\examples{
library(MASS)
result <- ryx(Boston, y = "medv")
print(result)

}
