% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Print a ryx object}
\usage{
\method{print}{ryx}(obj)
}
\arguments{
\item{obj}{A ryx object.}
}
\description{
Prints a ryx object in a nicely formatted and informative manner.
}
\examples{
library(MASS)
print(ryx(Boston, y="medv"))
}
