% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{print.ryx}
\usage{
\method{print}{ryx}(x)
}
\arguments{
\item{x:}{an object of class ryx}
}
\value{
Prints out an object displaying the variables chosen, the correlations of each variable, the p-value of each correlation, and the significance of each p-value
}
\description{
The print function prints out an object displaying the correlation, the p-values and the significance of the p-values of all the selected outcome variables against the response variable.
}
\examples{
\dontrun{
print(ryx(mtcars, y = "mpg", x = c("hp", "wt", "disp", "cyl", "am", "gear")))
}

}
