% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rxy.R
\name{print.ryx}
\alias{print.ryx}
\title{Printing RYX Function}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{the results from "rxy" function}
}
\value{
A formatted print of the results of the "rxy" function
}
\description{
Prints RXY function in a neat clean way
}
\examples{
library(MASS)
data(Boston)
print(ryx(Boston, y="medv"))
}
