% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Create an ryx object}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{datafile to import.}

\item{y}{dependent variable}

\item{x}{list of independent variables to compare to dependent variable
If not specified, will use all other variables}
}
\value{
a ryx object
}
\description{
This code will take a dataset and find the correlations between
a given dependent variable and each variable in a list of independent variables
}
\examples{
\dontrun{
data(Boston, package = "MASS")
x <- ryx(Boston, y = "medv")
}
}
