% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Compute and explore correlations between different variables
Compute, summarize and plot correlations between variables}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame}

\item{y}{A numeric variable}

\item{x}{One ore more numeric variables}
}
\value{
Results in a list
}
\description{
Compute and explore correlations between different variables
Compute, summarize and plot correlations between variables
}
\examples{
# Correlate mpg with the rest of the variables in mtcars
x <- ryx(mtcars, y = "mpg")
# Print results
print(x)
# Summarize results in words
summary(x)
# Plot results
plot(x)
}
