% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Calculate correlations between a dependent and multiple independent variables}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{y}{The dependent variable (character).}

\item{x}{The independent variables (character vector).}
}
\value{
An object of class `ryx` with results.
}
\description{
Calculate correlations between a dependent and multiple independent variables
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
print(x)
}
