% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Calculate correlations in a data frame}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame.}

\item{y}{The name of one of the variables in data.}

\item{x}{A character vector: the names of other variables in data.
If missing, the numeric variables in data other than y will be used.}
}
\value{
A list including y, x, and a data frame of results, which includes
correlation values, p-values, and significance values for each variable
in x when tested against y.
}
\description{
Calculates correlations between one variable (y) and all others (x)
in a data frame.
}
\examples{
library(MASS)
ryx(Boston, y="medv")
}
