% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{is a data frame}

\item{y}{is the name of a numeric variable in the data frame}

\item{x}{is a character vector with the names of one or more numeric variables in the data frame}
}
\value{
An object that displays the variables chosen, the correlations of each variable, the p-value of each correlation, and the significance of each p-value
}
\description{
Correlation Function
}
\details{
The ryx package intends to make calculating and visualizing correlations easier.
}
\examples{
\dontrun{
ryx(mtcars, y = "mpg", x = c("hp", "wt", "disp", "cyl", "am", "gear"))
}
}
