% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary method for ryx objects}
\usage{
\method{summary}{ryx}(object, ...)
}
\arguments{
\item{object}{An object of class `ryx`, containing the correlation results from the `ryx` function.}
}
\description{
Provides a textual summary of the correlations, including the median absolute correlation and the range.
}
\details{
The summary includes the median absolute correlation, the range of correlation values,
and the number of significant correlations at the p < 0.05 level.
}
\examples{
library(MASS)
result <- ryx(Boston, y = "medv")
summary(result)

}
