% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summarize a ryx object}
\usage{
\method{summary}{ryx}(obj)
}
\arguments{
\item{obj}{A ryx object.}
}
\description{
Summarizes a ryx object, providing statistical info about its correlations.
}
\examples{
library(MASS)
summary(ryx(Boston, y="medv"))
}
