% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{summary.ryx}
\usage{
\method{summary}{ryx}(x)
}
\arguments{
\item{x:}{an object of class ryx}
}
\value{
A summary of the statistics calculated in ryx.
}
\description{
The summary function summarized in only a few sentences what variables your a correlating with your response variable. It displays the median, the range, the amount of statistically significant correlations, and all the variables involved the variables used.
}
\examples{
\dontrun{
summary(ryx(mtcars, y = "mpg", x = c("hp", "wt", "disp", "cyl", "am", "gear")))
}
}
