% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rxy.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Written summary of results}
\usage{
\method{summary}{ryx}(object, ...)
}
\arguments{
\item{x}{the results from "rxy" function}
}
\value{
A formatted written summary of the results of the "rxy" function
}
\description{
Printing out a written summary of the results from the rxy function
}
\examples{
library(MASS)
data(Boston)
summary(ryx(Boston, y="medv"))
}
